revert changes for create docxf
