+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -121,12 +121,11 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
             fileName = request.getParameter("fileName");
             String fileExt = request.getParameter("fileExt");
             String pageID = request.getParameter("pageId");
-            String attachmentTemplateId = request.getParameter("attachmentTemplateId");
 
             if (pageID != null && !pageID.equals("")) {
                 try {
                     Long attachmentId = documentManager.createDemo(fileName, fileExt,
-                            Long.parseLong(pageID), convertManager.getMimeType(fileExt), attachmentTemplateId);
+                            Long.parseLong(pageID), convertManager.getMimeType(fileExt));
 
                     response.sendRedirect( request.getContextPath() +  "?attachmentId=" + URLEncoder.encode(attachmentId.toString(), "UTF-8"));
                     return;
