+++ src/main/java/onlyoffice/managers/convert/ConvertManagerImpl.java
@@ -101,27 +101,14 @@ public class ConvertManagerImpl implements ConvertManager {
     }
 
     public JSONObject convert(Long attachmentId, String key, String ext) throws Exception {
-        String url = urlManager.getFileUri(attachmentId);
-        String convertToExt = convertsTo(ext);
-
-        return convert(key, ext, convertToExt, url, true);
-    }
-
-    public JSONObject convert(Long attachmentId, String key, String currentExt, String convertToExt, boolean async) throws Exception {
-        String url = urlManager.getFileUri(attachmentId);
-
-        return convert(key, currentExt, convertToExt, url, async);
-    }
-
-    public JSONObject convert(String key, String currentExt, String convertToExt, String url, boolean async) throws Exception {
         CloseableHttpClient httpClient = HttpClients.createDefault();
         JSONObject body = new JSONObject();
-        body.put("async", async);
+        body.put("async", true);
         body.put("embeddedfonts", true);
-        body.put("filetype", currentExt);
-        body.put("outputtype", convertToExt);
+        body.put("filetype", ext);
+        body.put("outputtype", convertsTo(ext));
         body.put("key", key);
-        body.put("url", url);
+        body.put("url", urlManager.getFileUri(attachmentId));
 
         StringEntity requestEntity = new StringEntity(body.toString(), ContentType.APPLICATION_JSON);
         HttpPost request = new HttpPost(urlManager.getInnerDocEditorUrl()
